Split machine types to fix build.
